+++ src/main/java/mod/tjt01/lapislib/core/datagen/LapisLibLanguage.java
@@ -33,5 +33,7 @@ public class LapisLibLanguage extends LanguageProvider {
         this.add("lapislib.common.config.green", "Green: %s");
         this.add("lapislib.common.config.blue", "Blue: %s");
         this.add("lapislib.common.config.alpha", "Alpha: %s");
+
+        this.add("lapislib.common.config.open_picker", "Open color picker");
     }
 }