+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 ## [1.19.2-1.1.0]
 ### Added
 - Ported to 1.20.1
+- The colored part of color picker buttons can now be changed by resource packs
+
+### Fixed
+- Fixed missing lang entry for color picker button
 
 ## [1.19.2-1.1.0]
 ### Added